Fix failed VIES calls since Oct 28th 2021
VIES calls started failing with this gem around October 28th. We traced
the issue back to the SOAPAction key being included in the request
header. Here is a quick fix to exclude the offending header.

For search purposes, the error we got was:
"Unknown error: (soap:Server) The given SOAPAction checkVat does not match an operation."